Provides zprint action
While `cljfmt` is wide adopted, there is a little inconvenience: it must
be installed as a plugin on the project, requiring the installation of
dependencies to run.

An alternative is `zprint`, which is has a very convenient global CLI
alternative through `boot-fmt`. `zprint` has a lot of knobs to adjust on
the formatting, and using `boot-fmt` allows for formatting without
dependencies being installed.

This commit creates the new action providing `zprint` as a formatter.