Use cljfmt without installing projects dependency
When using `cljfmt` action, we ended up downloading all the dependencies
of the project, as it would be integrated as a `lein` plugin.

This was cumbersome, as it required exposing extra configuration to
access the project's dependencies as well. Ideally, to reduce the
adoption, it would be better to use `cljfmt` as a globally accessible
tool, without the need to access the project's dependencies from GitHub
Actions' execution environment.

There is a [previously undocumented feature](https://github.com/weavejester/cljfmt/pull/164)
on `cljfmt` which exposes it as a `deps-tools` executable.

This commit refactors the execution to run the formatter from a global
execution path. This means we don't need extra access to the project
dependencies.

The configuration to use the formatter don't live anymore inside
`lein`'s `project.clj`, and now is expected to live on their own
file inside the project, as it is documented.

We are also taking a chance to refactor a couple of things on the
`zprint` action, which is somewhat complementary to `cljfmt`.